Deduplicate string-interpolation helpers into a shared module
The PLY and Lark backends carried near-identical copies of the string-processing
helpers. Extract them into src/inmanta/parser/strings.py (range merging, {{var}}
interpolation, f-string expansion, variable-to-Reference conversion, and the
interpolation regex) and have both backends call the shared functions.

The functions were verified equivalent across backends; the differential parser
test confirms both backends still produce identical ASTs. Removes ~200 lines of
duplication and gives string handling a single source of truth.

The regex-token handlers stay per-backend (PLY builds the Regex in the lexer, Lark
in the transformer) and PLY's bare-string Reference("self") is left for a separate
change, since fixing it needs a namespace the PLY lexer does not have.
