+++ src/inmanta/parser/lark_parser.py
@@ -23,7 +23,6 @@ import hashlib
 import logging
 import os
 import re
-import string
 import sys
 import tempfile
 import warnings
@@ -66,7 +65,7 @@ from inmanta.ast.statements.generator import ConditionalExpression, Constructor,
 from inmanta.ast.variables import AttributeReference, Reference
 from inmanta.const import CF_CACHE_DIR
 from inmanta.execute.util import NoneValue
-from inmanta.parser import InvalidNamespaceAccess, ParserException, ParserWarning
+from inmanta.parser import InvalidNamespaceAccess, ParserException, ParserWarning, strings
 from inmanta.parser.cache import CacheManager
 from inmanta.parser.keywords import RESERVED_KEYWORDS
 
@@ -189,11 +188,6 @@ def attach_to_project(project_dir: str) -> None:
 def detach_from_project() -> None:
     cache_manager.detach_from_project()
 
-
-# ---- Format-string regex (same as PLY parser) ----
-
-_format_regex = r"""({{\s*([\.A-Za-z0-9_-]+)\s*}})"""
-_format_regex_compiled = re.compile(_format_regex, re.MULTILINE | re.DOTALL)
 
 # Set of reserved keywords – used to reject keywords used as identifiers.
 # Lark's contextual lexer matches keywords as ID when ID is valid in the grammar,
@@ -461,10 +455,6 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
         none.namespace = self.namespace
         none.lexpos = lexpos
         return none
-
-    def _expand_range(self, start: Range, end: Range) -> Range:
-        """Merge two ranges into one spanning from start to end."""
-        return Range(start.file, start.lnr, start.start_char, end.end_lnr, end.end_char)
 
     def _validate_id(self, token: Token) -> None:
         """Raise ParserException if an ID token holds a reserved keyword value.
@@ -1278,7 +1268,7 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
         loc = Location(self.file, token.line or 1)
         decoded = _safe_decode(content, "Invalid escape sequence in string.", loc)
         # Check for format strings in dict keys
-        match_obj = _format_regex_compiled.findall(decoded)
+        match_obj = strings.INTERPOLATION_REGEX.findall(decoded)
         if len(match_obj) != 0:
             ls = self._locatable(token)
             ls = LocatableString(decoded, ls.location, ls.lexpos, ls.namespace)
@@ -1360,7 +1350,7 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
         self._validate_id(id_token)
         id_ls = self._locatable(id_token)
         merged_value = f"{str(left)}::{str(id_ls)}"
-        r = self._expand_range(left.location, id_ls.location)
+        r = strings.expand_range(left.location, id_ls.location)
         return LocatableString(merged_value, r, id_ls.lexpos, self.namespace)
 
     def class_ref_cid(self, cid_token: Token) -> LocatableString:
@@ -1376,7 +1366,7 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
     def class_ref_ns(self, left: LocatableString, _sep: Token, cid_token: Token) -> LocatableString:
         cid_ls = self._locatable(cid_token)
         merged_value = f"{str(left)}::{str(cid_ls)}"
-        r = self._expand_range(left.location, cid_ls.location)
+        r = strings.expand_range(left.location, cid_ls.location)
         return LocatableString(merged_value, r, cid_ls.lexpos, self.namespace)
 
     def class_ref_err_dot(self, var_ref: object, cid_token: Token) -> NoReturn:
@@ -1392,7 +1382,7 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
 
         full_string = LocatableString(
             f"{var_str}.{cid_ls}",
-            self._expand_range(var_str.location, cid_ls.location),
+            strings.expand_range(var_str.location, cid_ls.location),
             var_str.lexpos,
             self.namespace,
         )
@@ -1438,7 +1428,7 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
         loc = Location(self.file, token.line or 1)
         decoded = _safe_decode(content, "Invalid escape sequence in string.", loc)
         ls = LocatableString(decoded, self._range(token), token.start_pos or 0, self.namespace)
-        result = _get_string_ast_node(ls, False)
+        result = strings.get_string_ast_node(ls, False)
         result.location = self._loc(token)
         result.namespace = self.namespace
         return result
@@ -1450,7 +1440,7 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
         loc = Location(self.file, token.line or 1)
         decoded = _safe_decode(content, "Invalid escape sequence in f-string.", loc)
         ls = LocatableString(decoded, self._range(token), token.start_pos or 0, self.namespace)
-        result = _process_fstring(ls)
+        result = strings.process_fstring(ls)
         result.location = self._range(token)
         result.namespace = self.namespace
         return result
@@ -1465,7 +1455,7 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
 
     def const_mls(self, token: Token) -> Union[Literal, StringFormat]:
         ls = self._process_mls(token)
-        result = _get_string_ast_node(ls, True)
+        result = strings.get_string_ast_node(ls, True)
         result.location = self._range(token)
         result.namespace = self.namespace
         return result
@@ -1488,130 +1478,6 @@ class InmantaTransformer(Transformer[Token, list[Statement]]):
 
     def constants(self, *items: ExpressionStatement) -> list[ExpressionStatement]:
         return list(items)
-
-
-# ---- String processing helpers ----
-
-
-def _get_string_ast_node(string_ast: LocatableString, mls: bool) -> Union[Literal, StringFormat]:
-    """Process a string for interpolation, expanding {{var}} references into a StringFormat."""
-    matches: list[re.Match[str]] = list(_format_regex_compiled.finditer(str(string_ast)))
-    if len(matches) == 0:
-        return Literal(str(string_ast))
-
-    start_lnr = string_ast.location.lnr
-    start_char_pos = string_ast.location.start_char
-    whole_string = str(string_ast)
-    mls_offset: int = 3 if mls else 1
-
-    def char_count_to_lnr_char(position: int) -> tuple[int, int]:
-        before = whole_string[0:position]
-        lines = before.count("\n")
-        if lines == 0:
-            return start_lnr, start_char_pos + position + mls_offset
-        else:
-            return start_lnr + lines, position - before.rindex("\n")
-
-    locatable_matches: list[tuple[str, LocatableString]] = []
-    for match in matches:
-        start_line, start_char = char_count_to_lnr_char(match.start(2))
-        end_line, end_char = char_count_to_lnr_char(match.end(2))
-        r: Range = Range(string_ast.location.file, start_line, start_char, end_line, end_char)
-        locatable_string = LocatableString(match[2], r, string_ast.lexpos, string_ast.namespace)
-        locatable_matches.append((match[1], locatable_string))
-
-    return StringFormat(str(string_ast), _convert_to_references(locatable_matches, string_ast.namespace))
-
-
-def _process_fstring(string_ast: LocatableString) -> Union[StringFormatV2, Literal]:
-    """Process an f-string for interpolation, expanding {field} references into a StringFormatV2."""
-    formatter = string.Formatter()
-    try:
-        parsed = list(formatter.parse(str(string_ast)))
-    except ValueError as e:
-        raise ParserException(string_ast.location, str(string_ast), f"Invalid f-string: {e}")
-
-    start_lnr = string_ast.location.lnr
-    start_char_pos = string_ast.location.start_char + 2  # f" or f' prefix
-
-    locatable_matches: list[tuple[str, LocatableString]] = []
-
-    def locate_match(match: tuple[str, Optional[str], Optional[str], Optional[str]], scp: int, end_char: int) -> None:
-        assert match[1]
-        r: Range = Range(string_ast.location.file, start_lnr, scp, start_lnr, end_char)
-        locatable_string = LocatableString(match[1], r, string_ast.lexpos, string_ast.namespace)
-        locatable_matches.append((match[1], locatable_string))
-
-    cur_pos = start_char_pos
-    for match in parsed:
-        if not match[1]:
-            break
-        lit_len = len(match[0])
-        field_len = len(match[1])
-        brack_len = 1 if field_len else 0
-        cur_pos += lit_len + brack_len
-        end_char = cur_pos + field_len
-        locate_match(match, cur_pos, end_char)
-        cur_pos += field_len
-
-        if match[2]:
-            cur_pos += 1
-            sub_parsed = formatter.parse(match[2])
-            for submatch in sub_parsed:
-                if not submatch[1]:
-                    break
-                slit_len = len(submatch[0])
-                sfield_len = len(submatch[1])
-                sbrack_len = 1 if sfield_len else 0
-                cur_pos += slit_len + sbrack_len
-                end_char = cur_pos + sfield_len
-                locate_match(submatch, cur_pos, end_char)
-                cur_pos += sfield_len + sbrack_len
-
-        cur_pos += brack_len
-
-    return StringFormatV2(str(string_ast), _convert_to_references(locatable_matches, string_ast.namespace))
-
-
-def _convert_to_references(
-    variables: Sequence[tuple[str, LocatableString]], namespace: Namespace
-) -> list[tuple["Reference", str]]:
-    """Convert variable name strings to References with proper location tracking."""
-
-    def normalize(variable: str, locatable: LocatableString, offset: int = 0) -> LocatableString:
-        start_char = locatable.location.start_char + offset
-        end_char = start_char + len(variable)
-        var_left_trim = variable.lstrip()
-        left_spaces = len(variable) - len(var_left_trim)
-        var_full_trim = var_left_trim.rstrip()
-        right_spaces = len(var_left_trim) - len(var_full_trim)
-        r: Range = Range(
-            locatable.location.file,
-            locatable.location.lnr,
-            start_char + left_spaces,
-            locatable.location.lnr,
-            end_char - right_spaces,
-        )
-        return LocatableString(var_full_trim, r, locatable.lexpos, locatable.namespace)
-
-    var_list: list[tuple[Reference, str]] = []
-    for match, var in variables:
-        var_name: str = str(var)
-        var_parts: list[str] = var_name.split(".")
-        ref_ls: LocatableString = normalize(var_parts[0], var)
-        ref = Reference(ref_ls)
-        ref.location = ref_ls.location
-        ref.namespace = namespace
-        if len(var_parts) > 1:
-            offset = len(var_parts[0]) + 1
-            for attr in var_parts[1:]:
-                attr_ls: LocatableString = normalize(attr, var, offset=offset)
-                ref = AttributeReference(ref, attr_ls)
-                ref.location = attr_ls.location
-                ref.namespace = namespace
-                offset += len(attr) + 1
-        var_list.append((ref, match))
-    return var_list
 
 
 # ---- Error handling ----
